examples: rename skew self adjoint --> self adjoint
