+++ examples/seismic/self_adjoint/__init__.py
