+++ examples/seismic/self_adjoint/utils.py
